feat: available zones & available regions
